# docs/apidocs/localapidoc.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is an api doc for `seda configurator` in sap cloud foundry.",
    "version": "0.0.1",
    "title": "Seda Configurator",
    "termsOfService": "",
    "contact": {
      "email": "gokulraj.ramdass@sap.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/oauth/token": {
      "post": {
        "tags": [
          "app-auth"
        ],
        "summary": "OAUTH Token URL",
        "description": "JWT Token generator for the application based on credentials( client id and clientsecret)",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "plain/text"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "clientid",
            "description": "client id from the application.",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "clientsecret",
            "description": "clientsecret from the application",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Success token generated !"
          },
          "405": {
            "description": "Invalid credentials"
          },
          "500": {
            "description": "Error Occurred during processing"
          }
        }
      }
    },
    "/parse": {
      "post": {
        "tags": [
          "app-parser"
        ],
        "summary": "Input Parser",
        "description": "Input Parser using Recast NLP along with configurations.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "description": "Bearer jwttoken",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Input Text source",
            "required": true,
            "schema": {
              "$ref": "#/definitions/textsource"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Success  !!"
          },
          "405": {
            "description": "Invalid Token"
          },
          "500": {
            "description": "Error Occurred during processing"
          }
        }
      }
    },
...
    "/apps/all": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "get all apps from the db",
        "operationId": "getallapps",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Success data retrived !!"
          },
          "500": {
            "description": "Error Occurred during processing"
          }
        }
      }
    },
...
    "/apps/{appid}/intents/all": {
      "get": {
        "tags": [
          "intents"
        ],
        "summary": "get all intents",
        "description": "get all intents for the given application ID.",
        "operationId": "getallintents",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "appid",
            "description": "id of the application",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success data fetched !"
          },
          "405": {
            "description": "Invalid Input or App ID"
          },
          "500": {
            "description": "Error Occurred during processing"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "textsource": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string",
          "example": "any text",
          "description": "request text from chatbots."
        }
      }
    },
...
  }
}